Test (#5)
* Con Injeccion en Program

* Fixed

Cambiada dependencias de Entity poblandolo con las clases,program injectando interfaces

* commands to use Migrations include

agregada la aclaracion de como efectuar la migracion y seed de la base de datos para cada integrante
