+++ AlkemyWallet.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.3.32901.215
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AlkemyWallet", "AlkemyWallet\AlkemyWallet.csproj", "{26BC9647-16BE-4E61-B631-FAFE2FF3CE6A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AlkemyWallet", "AlkemyWallet\AlkemyWallet.csproj", "{26BC9647-16BE-4E61-B631-FAFE2FF3CE6A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
