+++ AlkemyWallet/Core/Models/DTO/UserDTO.cs
@@ -1,5 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using AlkemyWallet.Entities;
 
 namespace AlkemyWallet.Core.Models
 {
