+++ AlkemyWallet/Entities/Login.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace AlkemyWallet.Core.Models.DTO
+namespace AlkemyWallet.Entities
 {
     public class Login
     {
