+++ AlkemyWallet/Repositories/Interfaces/IAccountRepository.cs
@@ -1,4 +1,4 @@
-﻿using AlkemyWallet.Core.Models;
+﻿using AlkemyWallet.Entities;
 
 namespace AlkemyWallet.Repositories.Interfaces
 {
