+++ AlkemyWallet/Repositories/Interfaces/ITransactionRepository.cs
@@ -1,8 +1,8 @@
-﻿using AlkemyWallet.Core.Models;
+﻿using AlkemyWallet.Entities;
 
 namespace AlkemyWallet.Repositories.Interfaces
 {
-   
+
     public interface ITransactionRepository : IRepositoryBase<Transaction>
     {
