+++ AlkemyWallet/Repositories/Interfaces/IUserRepository.cs
@@ -1,9 +1,9 @@
-﻿using AlkemyWallet.Core.Models;
+﻿using AlkemyWallet.Entities;
 
 namespace AlkemyWallet.Repositories.Interfaces
 {
-  
-        public interface IUserRepository : IRepositoryBase<User>
+
+    public interface IUserRepository : IRepositoryBase<User>
         {
        
         }
