+++ AlkemyWallet/Repositories/UserRepository.cs
@@ -1,9 +1,9 @@
-﻿using AlkemyWallet.Core.Models;
+﻿using AlkemyWallet.Entities;
 using AlkemyWallet.Repositories.Interfaces;
 
 namespace AlkemyWallet.Repositories
-{   
-        public class UserRepository : RepositoryBase<User>, IUserRepository
+{
+    public class UserRepository : RepositoryBase<User>, IUserRepository
         {
             public UserRepository(WalletDbContext context)
                 : base(context)
